+++ src/Components/Nav/Components/Gift.js
@@ -1,24 +1,12 @@
 import React, { Component } from "react";
-import { Giftdata } from "../Navdata/Giftdata";
-import { Link } from "react-router-dom";
-import Contents from "./Contents/Contents";
+import { Giftdata } from "../Navdata/Navdata";
+import Submenu from "./Contents/Submenu";
 
 export default class Gift extends Component {
   render() {
     return (
       <div className="Gift">
-        <ul className="subnavBox">
-          {Giftdata.map((menu) => {
-            return (
-              <li className="submenuBox" key={menu}>
-                <Link to="#" className="mainItem">
-                  <span>{menu.title}</span>
-                </Link>
-                <Contents contents={menu.content} />
-              </li>
-            );
-          })}
-        </ul>
+        <Submenu navData={Giftdata} />
       </div>
     );
   }
