+++ src/Components/Nav/Components/Living.js
@@ -1,24 +1,12 @@
 import React, { Component } from "react";
-import { Livingdata } from "../Navdata/Livingdata";
-import { Link } from "react-router-dom";
-import Contents from "./Contents/Contents";
+import { Livingdata } from "../Navdata/Navdata";
+import Submenu from "./Contents/Submenu";
 
 export default class Living extends Component {
   render() {
     return (
       <div className="Living">
-        <ul className="subnavBox">
-          {Livingdata.map((menu) => {
-            return (
-              <li className="submenuBox" key={menu}>
-                <Link to="#" className="mainItem">
-                  <span>{menu.title}</span>
-                </Link>
-                <Contents contents={menu.content} />
-              </li>
-            );
-          })}
-        </ul>
+        <Submenu navData={Livingdata} />
       </div>
     );
   }
