+++ src/Components/Nav/Nav.scss
@@ -1,12 +1,13 @@
-* {
+body {
   margin: 0;
-  box-sizing: border-box;
   header {
     display: flex;
     width: 100%;
+
     nav {
       position: fixed;
       width: 100%;
+
       .navWrapper {
         background-color: white;
         display: flex;
@@ -15,13 +16,34 @@
       .allmenuBtn {
         img {
           width: 25px;
-          margin: 35px 0 0 80px;
+          margin: 55px 0 0 80px;
         }
       }
       .logo {
         img {
-          margin: 35px 40px 0 40px;
+          margin: 30px 40px 0 40px;
           width: 150px;
+        }
+      }
+      .popinfo {
+        margin-top: -40px;
+        display: flex;
+        justify-content: flex-end;
+
+        div {
+          margin-right: 70px;
+          width: 150px;
+          height: 150px;
+          background-color: white;
+          border: solid 1px black;
+          display: flex;
+          flex-direction: column;
+          text-align: center;
+          text-decoration: none;
+          justify-content: space-around;
+          .linkDeco {
+            text-decoration: none;
+          }
         }
       }
       .navIcon {
@@ -30,12 +52,14 @@
         align-items: center;
         justify-content: flex-end;
         padding-right: 82px;
+
         img {
           width: 25px;
-          margin: 10px;
-        }
-        img:hover {
-          cursor: pointer;
+          margin: auto 10px;
+
+          &:hover {
+            cursor: poiner;
+          }
         }
       }
       .subnav {
@@ -43,12 +67,14 @@
         width: 100%;
         background-color: rgba($color: #ffffff, $alpha: 0.9);
         justify-content: space-evenly;
+
         .subnavBox {
           .eventMenu {
             width: 100%;
-            padding: 30px 200px 30px 200px;
+            padding: 20px 200px 30px 200px;
             display: flex;
             margin-left: 20px;
+
             span {
               padding-left: 30px;
             }
@@ -58,6 +84,7 @@
           display: flex;
           flex-direction: column;
           padding: 30px 50px 30px 50px;
+
           span {
             opacity: 1;
             width: 100%;
@@ -71,18 +98,22 @@
       }
       ul {
         display: flex;
+        margin: 0;
+
         .menuItem {
           display: flex;
           flex-direction: column;
           font-weight: bold;
           font-size: 18px;
           margin: 0 15px;
-          padding-bottom: 30px;
-          padding-top: 20px;
+          padding-bottom: 45px;
+          padding-top: 45px;
+
           &:hover {
             color: red;
             border-bottom: solid 3px red;
             cursor: pointer;
+
             small {
               opacity: 1;
             }
