--- src/Components/Nav/Navdata/Eventdata.js
@@ -1,5 +1,0 @@
-export const Eventdata = [
-  {
-    content: ["이벤트", "기획전", "아울렛"],
-  },
-];
