--- src/Components/Nav/Navdata/Giftdata.js
@@ -1,18 +1,0 @@
-export const Giftdata = [
-  {
-    title: "페이퍼",
-    content: ["카드", "포장지"],
-  },
-  {
-    title: "데코레이션",
-    content: ["태그"],
-  },
-  {
-    title: "상자",
-    content: ["봉투", "초콜릿 상자", "선물상자", "쇼핑백"],
-  },
-  {
-    title: "맞춤 선물",
-    content: ["취미", "생일/기념일"],
-  },
-];
