--- src/Components/Nav/Navdata/Livingdata.js
@@ -1,46 +1,0 @@
-export const Livingdata = [
-  {
-    title: "가구/인테리어",
-    content: [
-      "거실가구",
-      "주방가구",
-      "수납가구",
-      "침실가구",
-      "사무실가구",
-      "침구류",
-      "홈데코",
-    ],
-  },
-  {
-    title: "생활/건강",
-    content: ["청소/욕실", "마스크", "생활마카", "연필/샤프"],
-  },
-  {
-    title: "주방",
-    content: [
-      "식기류",
-      "보관용기",
-      "잔/컵",
-      "조리기구",
-      "물병/텀블러",
-      "주방잡화",
-      "드립/포트",
-    ],
-  },
-  { title: "디지털/가전", content: ["생활가전", "주방가전"] },
-  {
-    title: "수납",
-    content: [
-      "박스/보관",
-      "소품수납",
-      "여행/캠핑",
-      "텐트",
-      "의자",
-      "매트/침대",
-      "테이블",
-      "취사용품",
-    ],
-  },
-  { title: "하비", content: ["DIY/KIT", "아트/북", "잉크랩", "원데이클래스"] },
-  { title: "모나르떼", content: ["모나르떼"] },
-];
