--- src/Components/Nav/Navdata/Monamipetdata.js
@@ -1,26 +1,0 @@
-export const Monamipetdata = [
-  {
-    title: "사료",
-    content: ["건식사료", "습식사료", "퍼피", "시니어"],
-  },
-  {
-    title: "간식",
-
-    content: [
-      "동결건조간식",
-      "껌",
-      "저키/트릿",
-      "비스킷/소세지",
-      "캔/파우치",
-      "수제간식/영양제",
-    ],
-  },
-  {
-    title: "용품",
-    content: ["홈리빙", "배변", "그루밍", "외출", "놀이/훈련", "패션"],
-  },
-  {
-    title: "프리미엄 식품관",
-    content: ["K9", "맥아담스", "카나간", "헤일로", "오리젠", "윔지스"],
-  },
-];
