+++ src/Components/Nav/Navdata/Navdata.js
@@ -1,0 +1,146 @@
+export const Storypickdata = [
+  { title: "Stationery", content: ["필기", "취미"] },
+  {
+    title: "MonamiPet",
+    content: ["Puppy's Diary", "훈련", "노하우", "일상", "푸드"],
+  },
+  { title: "Living&Life", content: ["모나르떼", "인테리어", "DIY", "Gift"] },
+  { title: "Gift", content: ["특별한 선물", "교육", "미술"] },
+];
+
+export const Stationerydata = [
+  {
+    title: "프리미엄 펜",
+    content: [
+      "153프리미엄",
+      "데스크펜/스마트펜",
+      "파카",
+      "카렌다쉬",
+      "카웨코",
+      "매뉴스크립트",
+    ],
+  },
+  {
+    title: "펜/펜슬",
+    content: ["볼펜", "수성펜/사인펜", "연필/샤프", "형광펜"],
+  },
+  {
+    title: "마카",
+    content: ["네임펜/유성매직", "생활/보드마카", "패브릭/세라믹마카"],
+  },
+  {
+    title: "컬러링/브러쉬",
+    content: ["컬러링펜", "브러쉬펜", "미술용품"],
+  },
+  {
+    title: "디자인 문구",
+    content: [
+      "몰스킨",
+      "다이어리/노트",
+      "카드/봉투",
+      "파우치",
+      "풀/스티커/테이프",
+      "스탬프",
+      "데스크용품",
+      "모나미 MD상품",
+    ],
+  },
+];
+
+export const Monamipetdata = [
+  {
+    title: "사료",
+    content: ["건식사료", "습식사료", "퍼피", "시니어"],
+  },
+  {
+    title: "간식",
+    content: [
+      "동결건조간식",
+      "껌",
+      "저키/트릿",
+      "비스킷/소세지",
+      "캔/파우치",
+      "수제간식/영양제",
+    ],
+  },
+  {
+    title: "용품",
+    content: ["홈리빙", "배변", "그루밍", "외출", "놀이/훈련", "패션"],
+  },
+  {
+    title: "프리미엄 식품관",
+    content: ["K9", "맥아담스", "카나간", "헤일로", "오리젠", "윔지스"],
+  },
+];
+
+export const Livingdata = [
+  {
+    title: "가구/인테리어",
+    content: [
+      "거실가구",
+      "주방가구",
+      "수납가구",
+      "침실가구",
+      "사무실가구",
+      "침구류",
+      "홈데코",
+    ],
+  },
+  {
+    title: "생활/건강",
+    content: ["청소/욕실", "마스크", "생활마카", "연필/샤프"],
+  },
+  {
+    title: "주방",
+    content: [
+      "식기류",
+      "보관용기",
+      "잔/컵",
+      "조리기구",
+      "물병/텀블러",
+      "주방잡화",
+      "드립/포트",
+    ],
+  },
+  { title: "디지털/가전", content: ["생활가전", "주방가전"] },
+  {
+    title: "수납",
+    content: [
+      "박스/보관",
+      "소품수납",
+      "여행/캠핑",
+      "텐트",
+      "의자",
+      "매트/침대",
+      "테이블",
+      "취사용품",
+    ],
+  },
+  { title: "하비", content: ["DIY/KIT", "아트/북", "잉크랩", "원데이클래스"] },
+  { title: "모나르떼", content: ["모나르떼"] },
+];
+
+export const Giftdata = [
+  {
+    title: "페이퍼",
+    content: ["카드", "포장지"],
+  },
+  {
+    title: "데코레이션",
+    content: ["태그"],
+  },
+  {
+    title: "상자",
+    content: ["봉투", "초콜릿 상자", "선물상자", "쇼핑백"],
+  },
+  {
+    title: "맞춤 선물",
+    content: ["취미", "생일/기념일"],
+  },
+];
+
+export const Eventdata = [
+  {
+    content: ["이벤트", "기획전", "아울렛"],
+  },
+];
