--- src/Components/Nav/Navdata/Stationerydata.js
@@ -1,38 +1,0 @@
-export const Stationerydata = [
-  {
-    title: "프리미엄 펜",
-    content: [
-      "153프리미엄",
-      "데스크펜/스마트펜",
-      "파카",
-      "카렌다쉬",
-      "카웨코",
-      "매뉴스크립트",
-    ],
-  },
-  {
-    title: "펜/펜슬",
-    content: ["볼펜", "수성펜/사인펜", "연필/샤프", "형광펜"],
-  },
-  {
-    title: "마카",
-    content: ["네임펜/유성매직", "생활/보드마카", "패브릭/세라믹마카"],
-  },
-  {
-    title: "컬러링/브러쉬",
-    content: ["컬러링펜", "브러쉬펜", "미술용품"],
-  },
-  {
-    title: "디자인 문구",
-    content: [
-      "몰스킨",
-      "다이어리/노트",
-      "카드/봉투",
-      "파우치",
-      "풀/스티커/테이프",
-      "스탬프",
-      "데스크용품",
-      "모나미 MD상품",
-    ],
-  },
-];
