--- src/Components/Nav/Navdata/Storypickdata.js
@@ -1,9 +1,0 @@
-export const Storypickdata = [
-  { title: "Stationery", content: ["필기", "취미"] },
-  {
-    title: "MonamiPet",
-    content: ["Puppy's Diary", "훈련", "노하우", "일상", "푸드"],
-  },
-  { title: "Living&Life", content: ["모나르떼", "인테리어", "DIY", "Gift"] },
-  { title: "Gift", content: ["특별한 선물", "교육", "미술"] },
-];
