+++ src/Pages/Main/Main.js
@@ -1,0 +1,12 @@
+import React, { Component } from "react";
+import Nav from "../../Components/Nav/Nav";
+
+export default class Main extends Component {
+  render() {
+    return (
+      <div>
+        <Nav />
+      </div>
+    );
+  }
+}
