--- src/Styles/common.scss
@@ -1,11 +1,0 @@
-@mixin center {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-@mixin between {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
